+++ src/cpu/irq.c
@@ -1965,12 +1965,12 @@ static void context_init(exception_frame_t * __restrict oldframe, void * fn, voi
 
 static volatile uint8_t threads_not_started = 1;
 
-static uint_fast32_t get_td_us(uint_fast32_t timeUS)
+static uint_fast64_t get_td_us(uint_fast32_t timeUS)
 {
 	return timeUS * (cpu_getdebugticksfreq() / (1000 * 1000));
 }
 
-static uint_fast32_t get_td_ms(uint_fast32_t timeMS)
+static uint_fast64_t get_td_ms(uint_fast32_t timeMS)
 {
 	return timeMS * (cpu_getdebugticksfreq() / 1000);
 }
@@ -2519,10 +2519,10 @@ void local_delay_ms(uint_fast32_t timeMS)
 	if (threads_not_started)
 	{
 
-		const uint_fast32_t t0 = cpu_getdebugticks();
-		const uint_fast32_t td = get_td_ms(timeMS);
+		const uint_fast64_t t0 = cpu_getdebugticks();
+		const uint_fast64_t td = get_td_ms(timeMS);
 		//PRINTF("1 local_delay_ms: t0=0x%08X, td=0x%08X, irql=%u\n", (unsigned) t0, (unsigned) td, (unsigned) GIC_GetInterfacePriorityMask());
-		while ((uint32_t) (cpu_getdebugticks() - t0) < td)
+		while ((uint64_t) (cpu_getdebugticks() - t0) < td)
 			;
 	}
 	else
@@ -2545,13 +2545,13 @@ int local_wait8mask(volatile const uint8_t * flag, uint_fast8_t mask, uint_fast8
 		return 0;
 	if (threads_not_started)
 	{
-		const uint_fast32_t t0 = cpu_getdebugticks();
-		const uint_fast32_t td = get_td_ms(timeMS);
+		const uint_fast64_t t0 = cpu_getdebugticks();
+		const uint_fast64_t td = get_td_ms(timeMS);
 		do
 		{
 			if (((* flag & mask) == state))
 				return 0;
-		} while ((uint32_t) (cpu_getdebugticks() - t0) < td);
+		} while ((uint64_t) (cpu_getdebugticks() - t0) < td);
 		return 1;
 	}
 	else
@@ -2576,13 +2576,13 @@ int local_wait32mask(volatile const uint32_t * flag, uint_fast32_t mask, uint_fa
 		return 0;
 	if (threads_not_started)
 	{
-		const uint_fast32_t t0 = cpu_getdebugticks();
-		const uint_fast32_t td = get_td_ms(timeMS);
+		const uint_fast64_t t0 = cpu_getdebugticks();
+		const uint_fast64_t td = get_td_ms(timeMS);
 		do
 		{
 			if (((* flag & mask) == state))
 				return 0;
-		} while ((uint32_t) (cpu_getdebugticks() - t0) < td);
+		} while ((uint64_t) (cpu_getdebugticks() - t0) < td);
 		return 1;
 	}
 	else
@@ -2668,9 +2668,9 @@ void local_delay_ms(uint_fast32_t timeMS)
 	if (timeMS == 0)
 		return;
 
-    const uint_fast32_t t0 = cpu_getdebugticks();
-    const uint_fast32_t td = get_td_ms(timeMS);
-    while ((uint32_t) (cpu_getdebugticks() - t0) < td)
+    const uint_fast64_t t0 = cpu_getdebugticks();
+    const uint_fast64_t td = get_td_ms(timeMS);
+    while ((uint64_t) (cpu_getdebugticks() - t0) < td)
     	;
 }
 
@@ -2679,13 +2679,13 @@ void local_delay_ms(uint_fast32_t timeMS)
 // return non-zero: timeout error
 int local_wait8mask(volatile const uint8_t * flag, uint_fast8_t mask, uint_fast8_t state, uint_fast32_t timeMS)
 {
-	const uint_fast32_t t0 = cpu_getdebugticks();
-	const uint_fast32_t td = get_td_ms(timeMS);
+	const uint_fast64_t t0 = cpu_getdebugticks();
+	const uint_fast64_t td = get_td_ms(timeMS);
 	do
 	{
 		if (((* flag & mask) == state))
 			return 0;
-	} while ((uint32_t) (cpu_getdebugticks() - t0) < td);
+	} while ((uint64_t) (cpu_getdebugticks() - t0) < td);
 	return 1;
 }
 
@@ -2693,13 +2693,13 @@ int local_wait8mask(volatile const uint8_t * flag, uint_fast8_t mask, uint_fast8
 // return non-zero: timeout error
 int local_wait32mask(volatile const uint32_t * flag, uint_fast32_t mask, uint_fast32_t state, uint_fast32_t timeMS)
 {
-	const uint_fast32_t t0 = cpu_getdebugticks();
-	const uint_fast32_t td = get_td_ms(timeMS);
+	const uint_fast64_t t0 = cpu_getdebugticks();
+	const uint_fast64_t td = get_td_ms(timeMS);
 	do
 	{
 		if (((* flag & mask) == state))
 			return 0;
-	} while ((uint32_t) (cpu_getdebugticks() - t0) < td);
+	} while ((uint64_t) (cpu_getdebugticks() - t0) < td);
 	return 1;
 }
 // wait expected state of variable
@@ -2851,9 +2851,9 @@ void local_delay_us(uint_fast32_t timeUS)
 	if (timeUS == 0)
 		return;
 
-	const uint_fast32_t t0 = cpu_getdebugticks();	// Счетчик увеличивается с частотой процессора
-	const uint_fast32_t td = get_td_us(timeUS);
-	while ((uint32_t) (cpu_getdebugticks() - t0) < td)
+	const uint_fast64_t t0 = cpu_getdebugticks();	// Счетчик увеличивается с частотой процессора
+	const uint_fast64_t td = get_td_us(timeUS);
+	while ((uint64_t) (cpu_getdebugticks() - t0) < td)
 		;
 }
